論理削除実装

+++ app/photoGalleryModel.php
@@ -21,19 +21,21 @@ class photoGalleryModel extends Model
         $imagesAllPath = [];
         $search  = ["public", "\\"]; // パスを書き換えるのに利用
         $replace = ["storage", ""];  // パスを書き換えるのに利用
-        $imagesAll = photoGalleryModel::select('id', 'original_name', 'path')->get();
+        $imagesAll = photoGalleryModel::select('id', 'original_name', 'path', 'delete_flg')->get();
 
         // DBから取得したimageパスを配列に格納
         if (isset($imagesAll)) {
             foreach ($imagesAll as $image) {
-                $id = $image->id;
-                $name = $image->original_name;
-                $path = str_replace($search, $replace, $image->path);
-                $imagesAllPath[] = [
-                    'id'   => $id,
-                    'name' => $name,
-                    'path' => $path,
-                ];
+                if ($image->delete_flg === 1) {
+                    $id = $image->id;
+                    $name = $image->original_name;
+                    $path = str_replace($search, $replace, $image->path);
+                    $imagesAllPath[] = [
+                        'id'   => $id,
+                        'name' => $name,
+                        'path' => $path,
+                    ];
+                }
             }
         }
 
@@ -70,11 +72,21 @@ class photoGalleryModel extends Model
      */
     public function deleteImage(string $id)
     {
-        $image = photoGalleryModel::find($id);
-        Storage::delete($image->path);             // imageフォルダから画像を削除
-        DB::transaction(function () use ($image) { // DBから特定の画像データを削除
-            $image->delete();
-        });
+        $query = DB::table('user_setting')
+            ->select('delmode')
+            ->where('user_id', 1)
+            ->first();
+        if ($query->delmode === 1) {
+            DB::transaction(function () use ($id) {
+                photoGalleryModel::where('id', $id)->update(['delete_flg' => 0]);
+            });
+        } else {
+            $image = photoGalleryModel::find($id);
+            Storage::delete($image->path);             // imageフォルダから画像を削除
+            DB::transaction(function () use ($image) { // DBから特定の画像データを削除
+                $image->delete();
+            });
+        }
     }
 
     /**
@@ -85,12 +97,13 @@ class photoGalleryModel extends Model
     public function setting(array $setData)
     {
         // 論理削除 or 物理削除
-        if ($setData['delMethod']) {
-        } else {
-        }
+        DB::table('user_setting')->insert([
+            'user_id' => 1, // IDは仮で指定
+            'delmode' => $setData['delMode']
+        ]);
 
+        // 全レコード物理削除
         if (isset($setData['allDelete'])) {
-            // 全レコード物理削除
             Storage::deleteDirectory('public/photoGallery_images');
             DB::transaction(function () {
                 DB::table('images_path')->delete();
